% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.stlmfit.R
\name{AIC.stlmfit}
\alias{AIC.stlmfit}
\title{Extract the AIC from a \code{\link{stlmfit}} object for comparing models.}
\usage{
\method{AIC}{stlmfit}(object, ...)
}
\arguments{
\item{object}{a \code{\link{stlmfit}} object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The AIC value of the stlmfit object.  Here, AIC for restricted maximum likelihood (REML) is computed as 2 times the
negative log-likelihood plus 2 times the number of covariance model parameters. For REML, AIC should only be used to compare two models with the same fixed effects but different spatial covariance structures.
}
\description{
Extract the AIC from a \code{\link{stlmfit}} object for comparing models.
}
\examples{
stlmfit_obj <- stlmfit(formula = response_na ~ x, data = samp_data,
 xcoord = "xcoords", ycoord = "ycoords", tcoord = "times",
 cor_model_sp = "gaussian", cor_model_t = "exponential") 
AIC(stlmfit_obj)
}
