% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_r.R
\name{build_r}
\alias{build_r}
\title{Build Correlation Matrix}
\usage{
build_r(cov_type = "exponential", range, dist_mat)
}
\arguments{
\item{cov_type}{is a covariance model type (either "exponential" or "gaussian").}

\item{range}{is the range parameter.}

\item{dist_mat}{is the distance matrix.}
}
\value{
r, a correlation matrix.
}
\description{
Builds a correlation matrix from a covariance type, range, and distance matrix.
}
\examples{

dist_mat <- stats::dist(samp_data[ ,c("xcoords", "ycoords")],
 diag = TRUE, upper = TRUE) |> as.matrix()

build_r(cov_type = "exponential", range = 4, dist_mat = dist_mat)
build_r(cov_type = "gaussian", range = 0.1, dist_mat = dist_mat)   
}
