% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.predict.R
\name{fp.predict}
\alias{fp.predict}
\title{Finite Population Spatio-temporal Prediction}
\usage{
fp.predict(spatiotempdata)
}
\arguments{
\item{spatiotempdata}{is an object from the `give.nas` function with simulated counts, spatial coordinates, and time points.}
}
\value{
A list with (1) the data frame with the predicted site-by-site predictions, (2) the true total for the current year, (3) the estimated total for the current year, and (4) the prediction variance.
}
\description{
\code{fp.predict} predicts the population total for the current year of sampling from sampled sites in the current year as well as sampled sites in past years.
}
