% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give.nas.R
\name{give.nas}
\alias{give.nas}
\title{Select Observations that Have Missing Values}
\usage{
give.nas(sim_obj, nsamp, method)
}
\arguments{
\item{sim_obj}{is an object from \code{sim.spat.temp}}

\item{nsamp}{is the number of sites to be sampled}

\item{method}{is the how to select which sites will be sampled (for now, the only \code{method} is "Random"). In the future, method might be different if we know that some sites tend to get sampled over and over again through time.}
}
\value{
a data frame with the spatial coordinates, time points, counts, counts with missing values. an indicator for which sites were sampled (1 = sampled, 0 = unsampled), and an indicator for which sites are in the current smapling year (1 = current, 0 = past).
}
\description{
The purpose of \code{give.nas} select certain values in the data frame output of \code{sim.spat.temp} to be missing.
}
