% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2LL.spatiotemp.REML.R
\name{m2LL.spatiotemp.REML}
\alias{m2LL.spatiotemp.REML}
\title{Covariance Parameter Estimation Function through RestrictedMaximum Likelihood.}
\usage{
m2LL.spatiotemp.REML(theta, zcol, XDesign, xcoord, ycoord, timepoints,
  CorModel)
}
\arguments{
\item{theta}{is the parameter vector of (nugget, partialsill, range,
(vector of mean parameters), rho), where rho is the correlation for 
an AR(1) time series}

\item{zcol}{is the response vector of counts, with \code{NA} values 
for the missing values for sites that were not observed at a particular
time point.}

\item{XDesign}{is the design matrix containing the covariates used to predict animal or plant abundance (including a column of 1's for the intercept) for all
sites at all time points..}

\item{xcoord}{is a vector of the x spatial coordinates (in UTM)}

\item{ycoord}{is a vector of the y spatial coordinates (in UTM)}

\item{timepoints}{is a vector of timepoints for the model}

\item{CorModel}{is the geostatistical spatial correlation model to be used. See the \code{corModels} documentation for possible models to use.}
}
\value{
A numeric output of minus 2 times the restricted log likelihood to be minimized by `optim` to obtain spatial parameter estimates.
}
\description{
The primary purpose of \code{m2LL.spatiotemp.REML.R} is to estimate the spatial
covariance parameters and fixed effects through maximum likelihood.
}
