% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moose_14_18}
\alias{moose_14_18}
\title{TOC Moose Survey Data 2014 - 2018 Subset}
\format{
A data frame object including:
\describe{
  \item{Surveyyear}{Year of the survey}
  \item{ID}{Unique Site ID}
  \item{totalmoosena}{Total moose observed at the site, with NA for sites that were not sampled}
  \item{SurveyID}{ID of the survey}
  \item{SurveyName}{Name of the survey}
  \item{Stratname}{Name of the stratum, as a character}
  \item{Counted}{Whether or not the site was sampled}
  \item{AreaMi}{Area of the site, in square miles}
  \item{ELEV_MEAN}{Average elevation in the site}
  \item{stratfact}{strata, as a factor}
  \item{xcoords}{site x-coordinate, in TM}
  \item{ycoords}{site y-coordinate, in TM}
  \item{samp_frame}{whether or not the site was in the sampling frame for that year}
}

#' @examples
data(moose_14_18)
names(moose_14_18)
summary(moose_14_18)
}
\usage{
moose_14_18
}
\description{
A data set that can be used with the \code{FPSpatioTemp} package.
}
\keyword{datasets}
