% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moose_14_20}
\alias{moose_14_20}
\title{TOC Moose Survey Data 2014 - 2020 Subset}
\format{
A data frame object including:
\describe{
  \item{totalmoosena}{Total moose observed at the site, with NA for sites that were not sampled}
  \item{xcoords}{site x-coordinate, in TM}
  \item{ycoords}{site y-coordinate, in TM}
  \item{Surveyyear}{Year of the survey}
  \item{newstrat}{Stratification variable with the most recent stratification in 2020 for every year}
  \item{strat_origin}{Stratification variable with the original stratification for each year}
  \item{samp_frame}{whether or not the site was in the sampling frame for that year}
  \item{yearind}{whether or not the Surveyyear variable is equal to 2020}
}

#' @examples
data(moose_14_20)
names(moose_14_20)
summary(moose_14_20)
}
\usage{
moose_14_20
}
\description{
A data set that can be used with the \code{FPSpatioTemp} package.
}
\keyword{datasets}
