% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moose_vignette}
\alias{moose_vignette}
\title{TOC Moose Survey Data - Small}
\format{
A data frame object including:
\describe{
  \item{count}{Total moose observed at the site, with NA for sites that were not sampled}
  \item{xcoords}{site x-coordinate, in TM}
  \item{ycoords}{site y-coordinate, in TM}
  \item{year}{Year of the survey}
  \item{strata}{strata (\code{LOW} or \code{HIGH}), as a factor, corresponding to the original stratification in each year}
  \item{area_mi}{Area of the site, in square miles}
  \item{elev_mean}{Average elevation in the site}
  \item{samp_frame}{whether or not the site was in the sampling frame for that year}
}
}
\usage{
moose_vignette
}
\description{
A data set that can be used with the \code{FPSpatioTemp} package. The data set is a subset of \code{moose_complete}, to be used in the vignette.
}
\examples{
data(moose_vignette)
names(moose_vignette)
summary(moose_vignette)
}
\keyword{datasets}
