% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stlmfit.R
\name{plot.stlmfit}
\alias{plot.stlmfit}
\title{Plot a histogram of the normalized residuals and a normalized residuals vs. fitted values scatterplot}
\usage{
\method{plot}{stlmfit}(x, ...)
}
\arguments{
\item{x}{is an object of class \code{\link{stlmfit}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
plots of normalized residuals and normalized residuals vs. fitted values
}
\description{
Provides a quick way to assess model assumptions of normality and constant variance. Note that, when predicting a total, normality is only required when the number of sites is small.
}
\examples{
stlmfit_obj <- stlmfit(formula = response_na ~ x, data = samp_data,
 xcoord = "xcoords", ycoord = "ycoords", tcoord = "times")
plot(stlmfit_obj)
}
