% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_t}
\alias{plot_t}
\title{Temporal Plot}
\usage{
plot_t(xcoords, ycoords, times, response, alpha = 0.4)
}
\arguments{
\item{xcoords}{a vector of x-coordinates}

\item{ycoords}{a vector of y-coordinates}

\item{times}{a vector of times}

\item{response, }{a vector of the response variable}

\item{alpha}{a line transparency parameter}
}
\value{
a plot.
}
\description{
Produces a temporal plot of a response variable grouped by location
}
\examples{
sim_data <- sim_spatiotemp(nx = 6, ny = 5, ntime = 4, betavec = 3,
      sp_de = 0.5, sp_range = 4, sp_ie = 0.5,
      t_de = 0.5, t_range = 0.7, t_ie = 0.5,
      spt_ie = 0.5)
 df <- sim_data$out_df
 plot_t(df$xcoords, df$ycoords, df$times, df$response)
}
