% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.predict.stlmfit.R
\name{print.predict.stlmfit}
\alias{print.predict.stlmfit}
\title{Prints a short summary for the \code{\link{predict.stlmfit}()} function.}
\usage{
\method{print}{predict.stlmfit}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{is a prediction object generated from \code{\link{predict.stlmfit}()}}

\item{digits}{is the number of digits to be displayed in the
model output}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function uses the object that is output from \code{\link{predict.stlmfit}()} of class \code{predict.stlmfit}.
}
\examples{
stlmfit_obj <- stlmfit(formula = response_na ~ x, data = samp_data,
 xcoord = "xcoords", ycoord = "ycoords", tcoord = "times") 
pred_obj <- predict(object = stlmfit_obj, wtscol = "predwts")
print(pred_obj)
}
