% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stlmfit.R
\name{print.stlmfit}
\alias{print.stlmfit}
\title{Prints the fitted coefficient table of a spatio-temporal fitted linear model.}
\usage{
\method{print}{stlmfit}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{is an object generated from \code{\link{stlmfit}()}}

\item{digits}{is the number of digits to be displayed in the
model output}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list with \itemize{
  \item a table of fixed effects estimates and estimated spatiotemporal covariance estimates
}
}
\description{
This function uses the object that is output from \code{\link{stlmfit}()} of class \code{stlmfit}.
}
\examples{
stlmfit_obj <- stlmfit(formula = response_na ~ x, data = samp_data,
 xcoord = "xcoords",
ycoord = "ycoords", tcoord = "times") 
print(stlmfit_obj)
}
