% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stlmfit.R, R/print.summary.stlmfit.R
\name{print.summary.stlmfit}
\alias{print.summary.stlmfit}
\title{Prints the summary of a fitted spatio-temporal linear model.}
\usage{
\method{print}{summary.stlmfit}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

\method{print}{summary.stlmfit}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{is an summary object generated from \code{\link{summary.stlmfit}()}}

\item{digits}{is the number of digits to be displayed in the
model output}

\item{signif.stars}{is an option to show which predictors
are significant.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list with \itemize{
  \item model formula
  \item summary statistics for the residuals.
  \item a table of fixed effects estimates and associated standard errors.
  \item estimated spatial covariance parameter estimates.
}

a list with \itemize{
  \item model formula
  \item summary statistics for the residuals.
  \item a table of fixed effects estimates and associated standard errors.
  \item estimated spatial covariance parameter estimates.
}
}
\description{
This function uses the object that is output from \code{\link{summary.stlmfit}()}.

This function uses the object that is output from \code{\link{summary.stlmfit}()}.
}
\examples{
stlmfit_obj <- stlmfit(formula = response_na ~ x, data = samp_data,
 xcoord = "xcoords",
ycoord = "ycoords", tcoord = "times") 
print(summary(stlmfit_obj))
stlmfit_obj <- stlmfit(formula = response_na ~ x, data = samp_data,
 xcoord = "xcoords", ycoord = "ycoords", tcoord = "times") 
print(summary(stlmfit_obj))
}
