% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_moose}
\alias{raw_moose}
\title{TOC Raw Moose Survey Data 2004 - 2020}
\format{
A data frame object including:
\describe{
  \item{Surveyyear}{Year of the survey}
  \item{ID}{Unique Site ID}
  \item{SurveyID}{ID of the survey}
  \item{SurveyName}{Name of the survey}
  \item{Stratname}{Name of the stratum, as a character}
  \item{Counted}{Whether or not the site was sampled}
  \item{AreaMi}{Area of the site, in square miles}
  \item{ELEV_MEAN}{Average elevation in the site}
  \item{Intid}{}
  \item{GMU}{Game Unit Number}
  \item{centrlat}{latitude centroid}
  \item{centrlon}{longitude centroid}
  \item{totalmoose}{}
  \item{datecounted}{}
  \item{searchmin}{}
  \item{comments}{}
  \item{prp_burn_1_30}{}
  \item{prp_burn_31_60}{}
  \item{prop.cov.1}{}
  \item{prop.cov.2}{}
  \item{prop.cov.3}{}
  \item{prop.cov.4}{}
  \item{prop.cov.5}{}
  \item{prop.cov.6}{}
  \item{prop.cov.6_7}{}
  \item{prop.cov.7}{}
  \item{prop.cov.8}{}
  \item{prop.cov.9}{}
  \item{prop.cov.10}{}
  \item{prop.cov.11}{}
  \item{prop.cov.12}{}
  \item{prop.cov.13}{}
  \item{prop.cov.14}{}
  \item{prop.cov.15}{}
  \item{prop.cov.99}{}
  \item{Season}{}
  \item{Description}{}
  \item{Rand}{}
  \item{Firingorder}{}
  \item{Strat}{}
  \item{Stratmoose}{}
  \item{StratTracks}{}
  \item{Strathab}{}
  \item{Selected}{}
  \item{UnitID}{}
  \item{PERC_HABITAT}{}
}

#' @examples
data(raw_moose)
names(raw_moose)
summary(raw_moose)
}
\usage{
raw_moose
}
\description{
A data set that can be used with the \code{FPSpatioTemp} package.
}
\keyword{datasets}
