% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.stlmfit.R
\name{residuals.stlmfit}
\alias{residuals.stlmfit}
\title{Extract Model Residuals from an \code{stlmfit} object.}
\usage{
\method{residuals}{stlmfit}(object, type = "raw", ...)
}
\arguments{
\item{object}{a \code{stlmfit} object generated from the \code{\link{stlmfit}()} function.}

\item{type}{residual type: \code{"raw"} (the default) or \code{"normalized"}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a vector of residuals, consisting of each observed response minus the estimated mean
}
\description{
Extract Model Residuals from an \code{stlmfit} object.
}
\examples{
stlmfit_obj <- stlmfit(formula = response_na ~ x,
 data = samp_data,
 xcoord = "xcoords", ycoord = "ycoords", tcoord = "times") 
residuals(stlmfit_obj, type = "raw")
residuals(stlmfit_obj, type = "normalized")
}
