% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samp_data}
\alias{samp_data}
\title{Example Data}
\format{
A data frame object including:
\describe{
  \item{times}{a vector of time points}
  \item{xcoords}{a vector of x-coordinates}
  \item{ycoords}{a vector of y-coordinates}
  \item{response}{a response variable with no missing values}
  \item{response_na}{a response variable with missing values}
  \item{predwts}{a vector of prediction weights}
  \item{x}{a covariate}
  \item{area}{the area of the site}
  \item{x_fact}{a predictor that is of type \code{<fct>}}
  \item{x_char}{a predictor that is of type \code{<chr>}}
  \item{x_miss}{x with a few NA values}
  \item{x_fact_miss}{x_fact with a few NA valus}
  \item{xcoords_miss}{xcoords with a few NA values}
  \item{times_miss}{times with a few NA values}
}
}
\usage{
samp_data
}
\description{
A toy data set that is small in size used for many examples and tests.
}
\examples{
data(samp_data)
names(samp_data)
summary(samp_data)
}
\keyword{datasets}
