% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.spat.temp.R
\name{sim.spat.temp}
\alias{sim.spat.temp}
\title{Simulate Spatio-temporal Data.}
\usage{
sim.spat.temp(coords_times, spatvec, rho, betavec, XDesign)
}
\arguments{
\item{coords_times}{is a data frame with x-coordinates in the first column, y-coordinates in the second column, and time points in the third column}

\item{spatvec}{is the vector of spatial parameters (partial sill, range)}

\item{rho}{is the autocorrelation parameter for an AR(1) time series.}

\item{betavec}{is the vector of coefficients for fixed effects.}

\item{XDesign}{is the design matrix for the fixed effects.}
}
\value{
a list with (1) a data frame with the \code{coords_times} input binded together with the simulated counts, (2) the spatio-temporal covariance matrix
}
\description{
The purpose of \code{m2ll.spatiotemp.mL.R} is to simulate a set of spatio-temporal data for a separable model with some fixed spatial coordinates and a few fixed time points.
}
