% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_spatiotemp.R
\name{sim_spatiotemp}
\alias{sim_spatiotemp}
\title{Simulation of Spatiotemporal Data}
\usage{
sim_spatiotemp(
  nx = 10,
  ny = 10,
  ntime = 5,
  betavec = 0,
  XDesign = matrix(1, nrow = nx * ny * ntime),
  sigma_parsil_spat = 0.9,
  range = 5,
  sigma_nugget_spat = 0.1,
  sigma_parsil_time = 0.7,
  rho = 0.8,
  sigma_nugget_time = 0.3,
  sigma_nugget_spacetime = 0.4,
  seed = round(runif(1, min = 1, max = 1e+07))
)
}
\arguments{
\item{nx}{is the number of x coordinates for a spatial grid}

\item{ny}{is the number of y coordinates for a temporal grid}

\item{ntime}{is the number of time points}

\item{betavec}{is the parameter vector for the fixed effects}

\item{XDesign}{is the fixed effects design matrix}

\item{sigma_parsil_spat}{is the spatial partial sill}

\item{range}{is the spatial range for an exponential covariance}

\item{sigma_nugget_spat}{is the spatial nugget}

\item{sigma_parsil_time}{is the temporal partial sill}

\item{rho}{is the temporal autocorrelation}

\item{sigma_nugget_time}{is the temporal nugget}

\item{sigma_nugget_spacetime}{is the independent variance parameter}

\item{seed}{a seed}
}
\value{
a list with \itemize{
    \item a data frame \code{out_df} containing the \code{response}, spatial coordinates \code{xcoords} and \code{ycoords}, and time points \code{times}
    }
}
\description{
The primary purpose of \code{sim_spatiotemp.R} is to simulate
spatiotemporal data from (for now) a linear sum-with-error model
that has both spatial covariance, temporal covariance, and 
completely independent random error.
}
\examples{
sim_spatiotemp(nx = 6, ny = 5, ntime = 4, betavec = 3,
       sigma_parsil_spat = 0.5, range = 4, sigma_nugget_spat = 0.5,
       sigma_parsil_time = 0.5, rho = 0.7, sigma_nugget_time = 0.5,
       sigma_nugget_spacetime = 0.5)
}
