% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlmfit.R
\name{stlmfit}
\alias{stlmfit}
\title{Spatiotemporal Model Fit}
\usage{
stlmfit(
  formula,
  data,
  xcoord,
  ycoord,
  tcoord,
  cor_model_sp = "exponential",
  cor_model_t = "exponential"
)
}
\arguments{
\item{formula}{is an \code{R} linear model formula specifying the
response variable as well as covariates for predicting the response on the unsampled sites.}

\item{data}{is a data frame or tibble with the response column, the covariates to be used for the block kriging, the spatio-temporal coordinates for all of the sites.}

\item{xcoord}{is the name (as a string) of the column in the data frame with x coordinates or longitudinal coordinates.}

\item{ycoord}{is the name (as a string) of the column in the data frame with y coordinates or latitudinal coordinates.}

\item{tcoord}{is the name (as a string) of the column in the data frame with 
the time points.}

\item{cor_model_sp}{a correlation model for the spatial correlation. Options are \code{"exponential"} (default), \code{"gaussian"}, \code{"spherical"}, or \code{"tent"}. Specific parameterizations of each spatial covariance type are available in Details.}

\item{cor_model_t}{a correlation model for the temporal correlation. Options are \code{"exponential"} (default), \code{"gaussian"}, \code{"spherical"}, or \code{"tent"}. Specific parameterizations of each temporal covariance type are available in Details.}
}
\value{
a list with \itemize{
  \item the original data set, appended with a few variables
  \item the estimated spatiotemporal covariance matrix
  \item the design matrix for the fixed effects
  \item a vector of the sampled response values
  \item a list with the 8 spatio-temporal parameter
   estimates for the fixed effects, 
   the Exponential spatial covariance and AR(1) temporal
   covariance and a sum-with-error linear mixed model
    (vector of fixed effects parameter estimates, spatial
   partial sill, range, spatial nugget, temporal partial sill,
   autocorrelation parameter, temporal nugget, and spatiotemporal
   nugget).
 }
}
\description{
Estimates regression coefficients, spatial autocorrelation
parameters, and temporal autocorrelation parameters, given spatial
coordinates, time points, a model formula, and a data frame.
}
\details{
The parameterizations for the correlation functions available for \code{cor_model_sp} and \code{cor_model_t} for distance \eqn{h} and general range parameter $\eqn{\phi}$ are
\itemize{
  \item{exponential: }{\eqn{exp(-3 h / \phi)}} 
  \item{spherical: }{\eqn{(1 - 1.5(h / \phi) + 0.5(h / \phi)^3) * I(h <= \phi)}}
  \item{gaussian: }{\eqn{exp(- 3 (h / \phi) ^2 )}}
  \item{tent: }{{\eqn{(1 - h / \phi) * I(h <= \phi)}}}
}
}
\examples{
stlmfit_obj <- stlmfit(formula = response_na ~ x, data = samp_data,
 xcoord = "xcoords", ycoord = "ycoords", tcoord = "times") 
}
