% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.stlmfit.R
\name{summary.stlmfit}
\alias{summary.stlmfit}
\title{Summarizes a fitted spatio-temporal linear model.}
\usage{
\method{summary}{stlmfit}(object, ...)
}
\arguments{
\item{object}{is an object generated from \code{\link{stlmfit}()} of class \code{stlmfit}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list with \itemize{
  \item model formula
  \item a table of fixed effects estimates and associated standard errors
  \item estimated spatial covariance parameter estimates
  \item residuals
       }
}
\description{
In conjunction with \code{print.summary.slmfit()}, the output looks similar
to output from \code{R}'s standard \code{lm()} function.
}
\examples{
stlmfit_obj <- stlmfit(formula = response_na ~ x, data = samp_data,
 xcoord = "xcoords", ycoord = "ycoords", tcoord = "times") 
summary(stlmfit_obj)
}
