% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_z_mats.R
\name{build_z_mats}
\alias{build_z_mats}
\title{Build Random Effect Matrices}
\usage{
build_z_mats(data_ord)
}
\arguments{
\item{data_ord}{is the name of a data.frame or tibble where the data has already been ordered to space within time..}
}
\value{
a list with \itemize{
  \item the spatial Z matrix.
  \item the temporal Z matrix
  }
}
\description{
Create Z random effect matrices for spatial and temporal effects.
}
