% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_z_mats.R
\name{build_z_mats}
\alias{build_z_mats}
\title{Build Random Effect Matrices}
\usage{
build_z_mats(data_ord)
}
\arguments{
\item{data_ord}{is the name of a data.frame or tibble where the data has already been ordered to space within time. This data frame has columns \code{_spindex}, which gives the spatial, \code{_tindex}, which gives the temporal index, and \code{.observed}, a logical giving whether the spatial index was observed at the time index.}
}
\value{
a list with \itemize{
  \item the spatial Z matrix, with only rows corresponding to \code{.observed = TRUE}
  \item the temporal Z matrix, with only rows corresponding to \code{.observed = TRUE}
  }
}
\description{
Create Z random effect matrices for spatial and temporal effects.
}
