% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_spt.R
\name{order_spt}
\alias{order_spt}
\title{Order Spatiotemporal Data}
\usage{
order_spt(data, xcoord, ycoord, tcoord)
}
\arguments{
\item{data}{is the name of a data.frame or tibble.}

\item{xcoord}{is the name of the column with the spatial x-coordinates.}

\item{ycoord}{is the name of the column with the spatial y-coordinates.}

\item{tcoord}{is the name of the column with the temporal indeces.}
}
\value{
a list with \itemize{
  \item the ordered data set, with an added logical variable, .observed, that
  is TRUE if the row was observed in the original data set.
  \item the large spatial distance matrix
  \item the small spatial distance matrix
  \item the large temporal distance matrix
  \item the small temporal distance matrix
  }
}
\description{
Order a spatiotemporal data set by space within time.
}
\examples{
example_df <- tibble::tibble(times = c(1, 1, 1, 2, 2, 3, 3, 3, 4, 4, 4),
xcoords = c(1, 1, 2, 1, 1, 1, 1, 2, 1, 1, 2),
ycoords = c(1, 2, 1, 1, 2, 1, 2, 1, 1, 2, 1),
resp = c(rnorm(5, 10, 4), NA, NA, rnorm(4, 10, 4)))
data_unordered <- dplyr::sample_n(example_df, size = nrow(example_df))
data_ord <- order_spt(data = data_unordered, xcoord = xcoords, ycoord = ycoords, tcoord = times)$full_data
}
