% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_cov}
\alias{plot_cov}
\title{Covariance Plot}
\usage{
plot_cov(
  stlmfit_obj,
  sp_epstol = c(0.2, 4, 20, Inf),
  t_epstol = c(0.2, 2, 6),
  xaxis_var = "temporal",
  ...
)
}
\arguments{
\item{stlmfit_obj}{an object fit with \code{stlmfit()}}

\item{sp_epstol}{a range of values used for colours for various spatial distances}

\item{t_epstol}{a range of values used for colours for various temporal distances}

\item{xaxis_var}{is \code{"spatial"} if the spatial distances should go on the 
x-axis and \code{"temporal"} if the temporal distances should go on the x-axis}

\item{...}{extra options to be passed to `\code{ggplot}`}
}
\value{
a plot with covariance on the y-axis and temporal distance on the x-axis
coloured by various spatial distances in \code{sp_epstol}.
}
\description{
Produces a plot of the covariance with time distance on the x-axis for various
degrees of spatial distance.
}
\examples{
set.seed(07262022)
obj <- sim_spatiotemp(nx = 6, ny = 5, ntime = 4, betavec = 3,
      sp_de = 0.5, sp_range = 4, sp_ie = 0.5,
      t_de = 0.5, t_range = 0.7, t_ie = 0.5,
      spt_ie = 0.5)
      
samp_obj <- sample_spatiotemp(obj = obj, n = 70, samp_type = "random")
samp_data <- samp_obj$df_full
samp_data <- samp_data |>
 dplyr::mutate(predwts = dplyr::if_else(times == max(times),
  true = 1, false = 0))
samp_data$x <- rnorm(nrow(samp_data), 0, 1)
samp_data <- samp_data
stlmfit_obj <- stlmfit(formula = response_na ~ x, data = samp_data,
 xcoord = "xcoords", ycoord = "ycoords", tcoord = "times") 
 
plot_cov(stlmfit_obj = stlmfit_obj, sp_epstol = c(0.01, 0.1, 0.5, Inf),
t_epstol = c(0.2, 1, 4))
}
