% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_sp}
\alias{plot_sp}
\title{Spatial Plot}
\usage{
plot_sp(xcoords, ycoords, times, response)
}
\arguments{
\item{xcoords}{a vector of x-coordinates}

\item{ycoords}{a vector of y-coordinates}

\item{times}{a vector of times}

\item{response, }{a vector of the response variable}
}
\value{
a plot.
}
\description{
Produces a spatial plot of a response variable faceted by time.
}
\examples{
sim_data <- sim_spatiotemp(nx = 6, ny = 5, ntime = 4, betavec = 3,
      sp_de = 0.5, sp_range = 4, sp_ie = 0.5,
      t_de = 0.5, t_range = 0.7, t_ie = 0.5,
      spt_ie = 0.5)
 df <- sim_data$out_df
 plot_sp(df$xcoords, df$ycoords, df$times, df$response)
 
}
