% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_t}
\alias{plot_t}
\title{Temporal Plot}
\usage{
plot_t(xcoords, ycoords, times, response, alpha = 0.4)
}
\arguments{
\item{xcoords}{a vector of x-coordinates}

\item{ycoords}{a vector of y-coordinates}

\item{times}{a vector of times}

\item{response, }{a vector of the response variable}
}
\value{
a plot.
}
\description{
Produces a temporal plot of a response variable grouped by location
}
\examples{
sim_data <- sim_spatiotemp(nx = 6, ny = 5, ntime = 4, betavec = 3,
       sigma_parsil_spat = 0.5, range = 4, sigma_nugget_spat = 0.5,
       sigma_parsil_time = 1, rho = 0.99, sigma_nugget_time = 0,
       sigma_nugget_spacetime = 0.5)
 df <- sim_data$out_df
 plot_t(df$xcoords, df$ycoords, df$times, df$response)
}
