% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.stlmfit.R
\name{predict.stlmfit}
\alias{predict.stlmfit}
\title{Abundance Prediction}
\usage{
\method{predict}{stlmfit}(object, wtscol, pred_level = 0.9, ...)
}
\arguments{
\item{object}{is an object of class `stlmfit`}

\item{wtscol}{is the name of the column in the data frame with
the weights for which sites we want the prediction for.}

\item{pred_level}{is the level used for a prediction interval}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list with the following \itemize{
  \item the prediction of the total abundance 
  \item the prediction variance
  \item a 90% prediction interval lower bound
  \item a 90% prediction interval upper bound
  \item a vector of site-by-site predictions for the unsampled sites
  \item a vector of site-by-site prediction variances for the unsampled sites
  \item the original data set appended with predictions
  \item the prediction interval level
 }
}
\description{
Predicts the abundance given an object from \code{stlmfit()}.
}
\examples{
obj <- sim_spatiotemp(nx = 6, ny = 5, ntime = 4, betavec = 3,
      sigma_parsil_spat = 0.5, range = 4, sigma_nugget_spat = 0.5,
      sigma_parsil_time = 0.5, rho = 0.7, sigma_nugget_time = 0.5,
      sigma_nugget_spacetime = 0.5)
      
samp_obj <- sample_spatiotemp(obj = obj, n = 70, samp_type = "random")
samp_data <- samp_obj$df_full
samp_data <- samp_data \%>\% 
dplyr::mutate(predwts = dplyr::if_else(times == max(times),
 true = 1, false = 0))
stlmfit_obj <- stlmfit(formula = response_na ~ 1, data = samp_data,
 xcoordcol = "xcoords",
ycoordcol = "ycoords", tcol = "times") 
predict(object = stlmfit_obj, wtscol = "predwts")
}
