% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.predict.stlmfit.R
\name{print.predict.stlmfit}
\alias{print.predict.stlmfit}
\title{Prints a short summary for the \code{\link{predict.stlmfit}()} function.}
\usage{
\method{print}{predict.stlmfit}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{is a prediction object generated from \code{\link{predict.stlmfit}()}}

\item{digits}{is the number of digits to be displayed in the
model output}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function uses the object that is output from \code{\link{predict.stlmfit}()} of class \code{predict.stlmfit}.
}
\examples{
obj <- sim_spatiotemp(nx = 6, ny = 5, ntime = 4, betavec = 3,
      sigma_parsil_spat = 0.5, range = 4, sigma_nugget_spat = 0.5,
      sigma_parsil_time = 0.5, rho = 0.7, sigma_nugget_time = 0.5,
      sigma_nugget_spacetime = 0.5)
      
samp_obj <- sample_spatiotemp(obj = obj, n = 70, samp_type = "random")
samp_data <- samp_obj$df_full
samp_data <- samp_data |> 
dplyr::mutate(predwts = dplyr::if_else(times == max(times),
 true = 1, false = 0))
stlmfit_obj <- stlmfit(formula = response_na ~ 1, data = samp_data,
 xcoordcol = "xcoords",
ycoordcol = "ycoords", tcol = "times") 
pred_obj <- predict(object = stlmfit_obj, wtscol = "predwts")
print(pred_obj)
}
