% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_spatiotemp.R
\name{sim_spatiotemp}
\alias{sim_spatiotemp}
\title{Simulation of Spatiotemporal Data}
\usage{
sim_spatiotemp(
  resp_type = "normal",
  nx = 10,
  ny = 10,
  ntime = 5,
  betavec = 0,
  XDesign = matrix(1, nrow = nx * ny * ntime),
  sp_de = 0.9,
  sp_ie = 0.1,
  t_de = 0.7,
  t_ie = 0.3,
  spt_de = 0.3,
  spt_ie = 0.4,
  sp_range = sqrt(2)/3,
  t_range = 1/3,
  seed = round(runif(1, min = 1, max = 1e+07))
)
}
\arguments{
\item{resp_type}{is the distribution of the response variable (either
\code{"normal"} or \code{"lognormal"})}

\item{nx}{is the number of x coordinates for a spatial grid on the unit square}

\item{ny}{is the number of y coordinates for a spatial grid on the unit square}

\item{ntime}{is the number of equally spaced time points on the [0, 1] interval.}

\item{betavec}{is the parameter vector for the fixed effects}

\item{XDesign}{is the fixed effects design matrix}

\item{sp_de}{is the spatial partial sill}

\item{sp_ie}{is the spatial nugget}

\item{t_de}{is the temporal partial sill}

\item{t_ie}{is the temporal nugget}

\item{spt_de}{is the spatiotemporal partial sill}

\item{spt_ie}{is the independent variance parameter}

\item{sp_range}{is the spatial range for an exponential covariance}

\item{t_range}{is the temporal range for exponential covariance}

\item{seed}{a seed}
}
\value{
a list with \itemize{
    \item a data frame \code{out_df} containing the \code{response}, spatial coordinates \code{xcoords} and \code{ycoords}, and time points \code{times}
    }
}
\description{
The primary purpose of \code{sim_spatiotemp.R} is to simulate
spatiotemporal data from a product-sum model.
}
\examples{
sim_spatiotemp(nx = 6, ny = 5, ntime = 4, betavec = 3,
      sp_de = 0.5, sp_range = 4, sp_ie = 0.5,
      t_de = 0.5, t_range = 0.7, t_ie = 0.5,
      spt_ie = 0.5)
}
