% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlmfit.R
\name{stlmfit}
\alias{stlmfit}
\title{Spatio-Temporal Model Fit}
\usage{
stlmfit(
  formula,
  data,
  xcoordcol,
  ycoordcol,
  tcol,
  areacol = NULL,
  CorModel = "Exponential"
)
}
\arguments{
\item{formula}{is an \code{R} linear model formula specifying the
response variable as well as covariates for predicting the response on the unsampled sites.}

\item{data}{is a data frame or tibble with the response column, the covariates to be used for the block kriging, the spatio-temporal coordinates for all of the sites.}

\item{xcoordcol}{is the name of the column in the data frame with x coordinates or longitudinal coordinates.}

\item{ycoordcol}{is the name of the column in the data frame with y coordinates or latitudinal coordinates.}

\item{tcol}{is the name of the column in the data frame with 
the time points.}

\item{areacol}{is the name of the column in the data frame 
with the site areas
NOTE: package has not yet been tested for use of this argument.}

\item{CorModel}{a correlation model for the spatial correlation ("Exponential" is the only choice for now).}
}
\value{
a list with \itemize{
  \item the original data set, appended with a few variables
  \item the estimated spatiotemporal covariance matrix
  \item the design matrix for the fixed effects
  \item a vector of the sampled response values
  \item a list with the 8 spatio-temporal parameter
   estimates for the fixed effects, 
   the Exponential spatial covariance and AR(1) temporal
   covariance and a sum-with-error linear mixed model
    (vector of fixed effects parameter estimates, spatial
   partial sill, range, spatial nugget, temporal partial sill,
   autocorrelation parameter, temporal nugget, and spatiotemporal
   nugget).
 }
NOTE: This function will eventually be split into stlmfit() to fit
the model and estcov() as a helper in model fitting
}
\description{
Estimates regression coefficients, spatial autocorrelation
parameters, and temporal autocorrelation parameters, given spatial
coordinates, time points and a model formula.
NOTE: Current version has not been tested with any fixed effects
in the design matrix X
NOTE: Current version only does Maximum Likelihood, though REML 
will be added.
}
\examples{
obj <- sim_spatiotemp(nx = 6, ny = 5, ntime = 4, betavec = 3,
      sigma_parsil_spat = 0.5, range = 4, sigma_nugget_spat = 0.5,
      sigma_parsil_time = 0.5, rho = 0.7, sigma_nugget_time = 0.5,
      sigma_nugget_spacetime = 0.5)
      
samp_obj <- sample_spatiotemp(obj = obj, n = 70, samp_type = "random")
samp_data <- samp_obj$df_full
samp_data <- samp_data \%>\%
 dplyr::mutate(predwts = dplyr::if_else(times == max(times),
  true = 1, false = 0))
samp_data$x <- rnorm(nrow(samp_data), 0, 1)
stlmfit_obj <- stlmfit(formula = response_na ~ x, data = samp_data,
 xcoordcol = "xcoords",
ycoordcol = "ycoords", tcol = "times") 
}
