% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlmfit.R
\name{stlmfit}
\alias{stlmfit}
\title{Spatio-Temporal Model Fit}
\usage{
stlmfit(
  formula,
  data,
  xcoordcol,
  ycoordcol,
  tcol,
  wtscol,
  areacol = NULL,
  CorModel = "Exponential"
)
}
\arguments{
\item{formula}{is an \code{R} linear model formula specifying the
response variable as well as covariates for predicting the response on the unsampled sites.}

\item{data}{is a data frame or tibble with the response column, the covariates to be used for the block kriging, the spatio-temporal coordinates for all of the sites.}

\item{xcoordcol}{is the name of the column in the data frame with x coordinates or longitudinal coordinates.}

\item{ycoordcol}{is the name of the column in the data frame with y coordinates or latitudinal coordinates.}

\item{tcol}{is the name of the column in the data frame with 
the time points.}

\item{wtscol}{is the name of the column in the data frame with
the weights for which sites we want the prediction for.}

\item{areacol}{is the name of the column in the data frame 
with the site areas
NOTE: package has not yet been tested for use of this argument.}

\item{CorModel}{a correlation model for the spatial correlation ("Exponential" is the only choice for now).}
}
\value{
a list with \itemize{
  \item the prediction of the total abundance
  \item the prediction variance
  \item a 90% prediction interval lower bound
  \item a 90% prediction interval upper bound
  \item a vector of site-by-site predictions for the unsampled sites
  \item a vector of site-by-site prediction variances for the unsampled sites
  \item a list with the 7 spatio-temporal covariance parameter
   estimates for Exponential spatial covariance and AR(1) temporal
   covariance and a sum-with-error linear mixed model (spatial
   partial sill, range, spatial nugget, temporal partial sill,
   autocorrelation parameter, temporal nugget, and spatiotemporal
   nugget).
 }
NOTE: This function will eventually be split into stlmfit() to fit
the model, estcov() as a helper in model fitting, 
and predict.stlmfit(), in much the same way that sptotal is split.
}
\description{
Estimates regression coefficients, spatial autocorrelation
parameters, and temporal autocorrelation parameters, given spatial
coordinates, time points and a model formula.
NOTE: Current version has not been tested with any fixed effects
in the design matrix X
NOTE: Current version only does Maximum Likelihood, though REML 
will be added.
}
